add DispatchGroup
Add DispatchGroup for parsing VKNews
